requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
boto3==1.24.67
botocore==1.27.67
click==8.1.3
Django==4.1
django-baton==2.4.1
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
flake8==5.0.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.15.0
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.23.2
pandas==1.4.4
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
Wand==0.6.10
